fix: Dashboard.tsx Cleanup
